+++ .github/workflows/api-tests.yml
@@ -1,4 +1,4 @@
-name: Films API Tests
+name: ShareIt API Tests
 
 on:
   pull_request:
@@ -35,7 +35,6 @@ jobs:
         run: |
           chmod a+x ./tests/.github/workflows/build.sh
           ./tests/.github/workflows/build.sh
-
       - name: Install Node
         uses: actions/setup-node@v1
         with:
@@ -44,18 +43,14 @@ jobs:
       - name: Install newman
         run: |
           npm install -g newman
-
       - name: Run Application
         run: |
           chmod a+x ./tests/.github/workflows/run.sh
           ./tests/.github/workflows/run.sh
-
-
       - name: Waiting for application is up
         run: |
           /tmp/wait-for-it.sh -t 60 localhost:8080
           echo "Application is up"
-
       - name: Run POSTMAN tests
         run: |
           newman run ./tests/postman/sprint.json --delay-request 50 -r cli
