+++ src/main/java/ru/practicum/shareit/request/ItemRequestDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.requests;
+package ru.practicum.shareit.request;
 
 import lombok.Builder;
 import lombok.Data;
