+++ src/main/java/ru/practicum/shareit/request/ItemRequestRepository.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.requests;
+package ru.practicum.shareit.request;
 
 import org.springframework.data.jpa.repository.JpaRepository;
