+++ src/main/java/ru/practicum/shareit/request/ItemRequestService.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.requests;
+package ru.practicum.shareit.request;
 
 import java.util.List;
