+++ src/main/resources/application.properties
@@ -1,8 +1,12 @@
 spring.jpa.hibernate.ddl-auto=none
-spring.jpa.hibernate.dialect=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.show_sql=true
+
 spring.sql.init.mode=always
+
 # TODO Append connection to DB
+
+
 logging.level.org.springframework.orm.jpa=INFO
 logging.level.org.springframework.transaction=INFO
 logging.level.org.springframework.transaction.interceptor=TRACE
@@ -10,7 +14,10 @@ logging.level.org.springframework.orm.jpa.JpaTransactionManager=DEBUG
 #---
 spring.config.activate.on-profile=ci,test
 spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.url=jdbc:h2:file:./db/shareit;;DB_CLOSE_ON_EXIT=TRUE
+spring.datasource.url=jdbc:h2:mem:shareit
 spring.datasource.username=test
 spring.datasource.password=test
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+
+
 spring.h2.console.enabled=true
